+++ modules/apim-adapter/src/main/java/com/axway/apim/adapter/apis/OrgFilter.java
@@ -3,7 +3,7 @@ package com.axway.apim.adapter.apis;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.http.NameValuePair;
 import org.apache.http.message.BasicNameValuePair;