+++ modules/apim-adapter/src/main/java/com/axway/apim/adapter/jackson/StateSerializer.java
@@ -1,6 +1,6 @@
 package com.axway.apim.adapter.jackson;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import com.axway.apim.api.API;
 import com.fasterxml.jackson.core.JsonGenerator;