+++ modules/apim-adapter/src/main/java/com/axway/apim/api/model/CorsProfile.java
@@ -2,7 +2,7 @@ package com.axway.apim.api.model;
 
 import java.util.Arrays;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 public class CorsProfile {