+++ modules/apis/src/main/java/com/axway/apim/apiimport/APIImportConfigAdapter.java
@@ -31,7 +31,7 @@ import java.util.ListIterator;
 import java.util.Map;
 
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.StringSubstitutor;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;